+++ packages/ui/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ## [unreleased]
 
+### Changed
+
+- `ui/VirtualList.VirtualListBasic` scroll animation method for `scrollMode: native`
+
 ### Fixed
 
 - `ui/Scroller` jumping back when scrolled by long press